Add `addUsername` method to `user` (#10480)
Similar to other changes we've made to add effectful methods to this `user` object. It means that these are passed down from the top of the application, rather than imported further down the tree. This will make it easier to swap between side effects on different platforms (apps, web, etc.).

Co-authored-by: Ioanna Kokkini <ioannakok@users.noreply.github.com>