Move format parsing into DCAR
Moves as much of the parsing of `ArticleFormat` as possible into DCAR, so we no longer have to rely on the CAPI client to do this. The current logic duplicates exactly what the CAPI client does, so that we don't change too much at once. Improvements to the parsing logic can be made in a future change, once we've confirmed everything works as before.

At the moment, there isn't enough of the CAPI data available from frontend to fully parse `ArticleFormat`. Three further pieces of information are needed:

1. Content type (`content.type` in CAPI)
2. Display hint (`content.fields.displayHint` in CAPI)
3. Whether a blog is currently live (`content.fields.liveBloggingNow` in CAPI)

In cases where these items are missing, the parsing logic falls back to the frontend definition of `ArticleFormat`. Once they are made available, in a future change, these fallbacks can be removed.

This change also renames `format.ts` to `articleFormat.ts`, to more closely resemble the name of the type.