Stop Auto-generating Apps Layout Stories (#10608)
* Stopped auto-generating apps layout stories

This is the second part of this work, with the first updating the web layout stories. Given that we no longer create layout stories for every possible format combination, it's simpler to specify the stories manually. This also removes certain assumptions about `ArticleFormat`, which will make it easier to update the way we handle it in some upcoming changes.

This change:

- Replaces the generated apps layout stories with manually specified layout stories.
- Removes the generation code for apps layout stories.
- Removes the `HydratedLayoutWrapper` file, as it's no longer needed.
- Updates the `HydratedLayout` decorator to be able to handle both web and apps, and dark mode.
- Updates docs and logs.
- Reintroduces the delay for onwards upper for both web and apps.