Exclude Storybook From Webpack Dependency Group (#10682)
Dependabot is not including the `react-webpack5` dependency in the storybook group. It is instead including it in the `webpack` dependency group.

This is not meant to happen. Dependabot is meant to include a dependency in the first group it matches with, and storybook is higher in the list.

This change is an experiment to see if excluding storybook dependencies from the webpack group fixes the problem.