Changed the order of article example links (#10979)
The assumption is that on the dev server the links clicked most frequently are the DEV links, rather than the CODE or PROD links. Therefore, this change moves those links to the left column, to make them quicker to find and click. The PROD links, as a result, move to the right column.