Add `TimelineAtomBlockElement` (#10996)
This was previously called `TimelineBlockElement`, but our convention is that we use the word `Atom` in the name of `Atom` types. This also allows us to repurpose `TimelineBlockElement` for the new timeline element.

For now this actually duplicates the element, because we have to synchronise changes in frontend. This will be repurposed in a future change.