Add a model for timeline elements
First, renamed the `TimelineEvent` type used by timeline atoms to `TimelineAtomEvent` to avoid confusion.

Then removed the now-unused `TimelineBlockElement` and replaced it with `FETimelineBlockElement`, to represent the new element coming from frontend, and `DCRTimelineBlockElement`, to represent the parsed/enhanced version in DCR. The DCR version is stricter than the one coming from frontend, enforcing the following rules:

- Timeline events must have a date
- Timeline sections must have a title
- Timelines with one section should be flattened into a single list of events

Finally, added an enhancer module for timelines to convert from the frontend type to the DCR one.