Add grid decorators
Some components are written with assumptions about the context in which they will appear. One such assumption is that they will appear in a page that has been laid out using the Guardian's grid.

For example, from the `leftCol` breakpoint thumbnail images will move themselves out of the flow of text and into the left column of the Guardian grid. However, in storybook there is no left column, so these images disappear outside the viewport of the story.

Storybook has a solution for when components need some surrounding context to render correctly: decorators. We already rely on this to set up the palette variables correctly, so that stories have the correct colours in light and dark mode.

This change introduces a set of "grid" decorators, which can be used to lay out components within stories using the Guardian grid. Not all components need to be wrapped in this way, so these are designed to be applied on a per-story or per-component basis, rather than globally as with the colour scheme decorators.

Returning to the example above, using one of these decorators (e.g. `centreColumnDecorator`) allows text and images to be placed on the Guardian grid as they would in a full article. This means the left column is then available within the story for thumbnail images to appear in at appropriate breakpoints.