Update the grid decorators to handle `globalColourScheme`
Previously the grid decorators did not work well with the `splitTheme` decorator. The `globalColourScheme` toolbar item is now available, which allows for a different way to specify multiple formats and colour schemes.

For now these grid decorators work with "light", "dark" and "vertical". They are turned off whenever "horizontal" is set, as they do not currently make sense in that context.

This also updates the `KeyTakeaways` and `QAndAExplainers` stories to work with the new toolbar item.