Storybook Toolbar Item To Set Colour Scheme (#11139)
With the introduction of dark mode we've found it useful to apply colour schemes and split layouts to stories via decorators: `lightDecorator`, `darkDecorator` and `splitTheme`. This change builds on that approach by allowing you to switch any story between one of four colour schemes via the storybook toolbar:

1. Light mode.
2. Dark mode.
3. Horizontal Split: display the story twice, on the left in light mode and on the right in dark mode.
4. Vertical Split: display the story twice, on the top in light mode and on the bottom in dark mode.

More details on exactly how this works can be found in the JSDoc comments in `globalColourScheme.ts`.

There are a few advantages to this approach. Firstly, a story no longer needs to be hardcoded to a given colour scheme or split theme layout; any story can be viewed in any one of the four variants. Secondly, and relatedly, we can use Chromatic's "modes" to decide on one or more colour schemes to snapshot a story with, without having to write multiple stories, and tightly couple our stories to the snapshot variations we want. Thirdly, it makes it clearer which format is being applied to a given story. Previously, if a story took `format` as an `arg`, that was always ignored in favour of the `defaultFormat` or a format applied in one of the colour scheme decorators. Now the story's `format` will be used when available, unless you specifically want to display multiple format variants using `parameters.formats`.

This change also updates the `AudioAtom` stories to take advantage of this new approach. In addition, it includes a small refactor to `preview.ts`, to make it clearer and remove the unused `viewports` export.

In addition, this adds Chromatic modes for `globalColourScheme`. This allows us to specify which colour schemes we'd like to snapshot for each story. This change applies one of these modes to the `AudioAtom` stories.