`Accordion` stories use colour scheme and modes
Updated the accordion stories to use the new colour scheme toolbar item instead of the `splitTheme` decorator. Alongside this, switched to using Chromatic's modes, which are intended to replace the `viewports` feature, and also allow combinations of colour schemes and viewports to be snapshotted.

In addition, made some small changes to the way the story is presented. It now uses a narrow viewport by default, where the accordion UI will be visible. It also shows the content on wider breakpoints too, as that is part of the component's intended behaviour.