Split `FEFrontCard` into a union type
Frontend represents this with a `trait` and case classes, i.e. a union type. It also sends a `type` field to DCAR to represent each member of the union, so we can replicate that in the DCAR types.

We want to split it up because it's a more accurate type, but also to make it easier to split downstream types in DCAR, e.g. `DCRFrontCard`. Ultimately this will make it easier to handle new kinds of cards, such as election trackers.

Co-authored-by: Ioanna Kokkini <ioannakok@hotmail.com>