Override showVideo toggle behaviour when we should not show an image
In dynamic/package, we don't show the card image in certain layouts and slots. For example, when we have four cards we should only be showing images in the first two cards. However, this rule is not respected if any of the 3rd or 4th cards has the showVideo toggle on. In this commit we override the showVideo flag behaviour so that we don't show the image when we shouldn't.