Update `AdSlot.apps` stories to Component Story Format v3
This is the default for Storybook 7+, and integrates better with some of its features.

Replaced the wrapper decorator with the right column decorator to better reflect where this slot will appear at different breakpoints. Switched from the legacy Chromatic `viewports` API to the new `modes` API. Updated the default viewport to a slightly larger size. Switched to the currently recommended way of specifying actions, with the `storybook/test` `fn`. Stopped supplying the `ref` arg to stories as it wasn't being used.