Update `ArticleHeadline` stories to Component Story Format v3
This is the default for Storybook 7+, and integrates better with some of its features.

Replaced the theme generators with our existing `getAllThemes` function. Removed the `Section`/`Flex`/`ArticleContainer` wrappers, as these don't wrap the component when it appears in the page. Added the centre column decorator instead. Removed decorators that add a main media, as these aren't needed to show the headline component specifically.

Added a function to display the `ArticleFormat` type as a `string`, and used it to populate the headline copy.