Only log a problem if the video id is missing (#11820)
This was always logging, regardless of whether the id was found.

Also switches from ERROR to WARN. We still render the page if this occurs, the video just isn't rendered. Therefore this could be a warning, which would mean it generates less noise in the logs, particularly when looking for other errors.