Correctly target ULs in TextBlockComponent
When the css attribute is used with Emotion a class is generated that contains the styles passed to that attribute. That class is applied to the element in question. Therefore, if a tag selector is used it will select tags of that type that are children of the element.

When the element in question is a ul element and the selector used is `ul > li` the `ul` part of it is targeting ULs that are children of the element. Instead, to target the `ul` itself you need to use `&:is(ul)`.

Co-authored-by: Jamie B <53781962+JamieB-gu@users.noreply.github.com>