Fix container button wrapping (#12122)
The `text-wrap` property is only available on recent browsers, which means this button doesn't wrap on browsers that don't support it. This causes it to run off the edge of the screen when it contains a lot of copy.

`white-space` can be used as a shorthand for both `text-wrap` and `white-space-collapse`, but is supported by much older browsers. `normal` here applies `text-wrap: wrap` and `white-space-collapse: collapse` (this is the collapse behaviour already being used here, so no change).