Update `ArticleTitle` stories to Component Story Format v3
This is the default for Storybook 7+, and integrates better with some of its features.

Removed the `Wrapper` and replaced it with the `leftColumnDecorator` applied to all stories. Set all Chromatic snapshots to use `splitHorizontal`, to capture light and dark modes simultaneously. Removed some stories that duplicated the functionality of other stories. Used the new `colourSchemeBackground` and `colourSchemeTextColour` parameters rather than relying on custom story wrapping code. Fixed a bug where the `sectionUrl` field contained an extra `/`.

Removed some props from the stories that were previously dropped from the component. Adding type-safety with CSFv3 allows these issues to be surfaced by the compiler.