Remove `ArticleDesign.PrintShop`
We no longer intend to handle this as a separate `ArticleDesign`. Anything that was previously `PrintShop` will now be one of the other `ArticleDesign`s.

Removed print shop parsing, fixtures, and stories. Also removed the corresponding frontend type, which frontend will no longer be sending to DCAR.

Using a pre-release version for now.