Move And Rename `DCRArticle` (#12279)
Moves the type into its own file, as it's not a part of the `frontend` model. Combined this file with the existing `article.ts` file, to bring type and functions together. Removed the `DCR` prefix, because a) the project is DCAR, and b) we're in the project so we don't need to qualify the name. `FEArticleType` remains to make it clear that type comes from frontend.

Co-authored-by: Ravi <7014230+arelra@users.noreply.github.com>