Create a new `Article` type (#12316)
In the early days of DCAR (DCR as it was then called), it was expected that frontend would be responsible for building the model of an article. DCAR would simply be the "rendering" layer, taking that model and passing it through a series of React components. As a result, the frontend model became the core article data type used throughout DCAR: `FEArticleType` (originally it had a different name, but this is the present-day equivalent).

However, as the project developed there was an increasing need for DCAR to alter this data type, essentially augmenting the frontend data model with changes of its own. As these changes were initially small, and because this model was widely used, the decision was made to "enhance" the model rather than replace it entirely. The "enhancers" made small changes to the data, whilst keeping the overall shape of the model intact.

Over time further enhancers were introduced and, in some cases, added new properties that were created entirely within DCAR and which frontend didn't know anything about. Each change took the model further away from the original shape of the frontend data, making the concept of a single `FEArticleType` increasingly inaccurate.

At the same time, these enhancers were limited in how much they could update this model to fit the increasingly complex DCAR codebase. They were, for the most part, constrained to the shape of a model that was designed to suit frontend and the necessity of sending the data as JSON over a network.

These problems were recognised some time ago, and when it came to migrating fronts to DCAR an alternate approach was adopted. For fronts, frontend sends through an `FEFrontType`, but this is parsed into a separate `DCRFrontType` early on in DCAR. This new model is then used throughout the rest of the codebase, and is able to take a shape that meets the requirements of DCAR.

Following on from this approach, this change introduces the same concept for articles. It decouples the DCAR version of the article model from the frontend version. `Article` will be the new, DCAR article model, and `FEArticleType` (likely renamed as `FEArticle`) will continue to be the frontend one. `FEArticleType` will be parsed into `Article` at an early stage, and `Article` is then free to adapt to the needs of DCAR, independent from the requirements of frontend and the JSON request.

As the article model is currently used in a large number of places, this change does not attempt to update everything at once. It switches over to the new `Article` type in a few key places, and renames the existing, frontend-based model to `ArticleDeprecated`.

The new `Article` model at first contains a single field, containing `ArticleDeprecated`. This allows a series of small changes to be made over time, updating the `Article` model and expanding its usage, whilst reducing usage of `ArticleDeprecated`. The plan is for `ArticleDeprecated` eventually to disappear, and `FEArticleType` to be used only for parsing into `Article`.