Transparent Background Only On Interactive Articles (#12356)
* Transparent background only on interactive articles

A transparent background was set on the body element to support interactives. However, it means that on the iOS app the background appears grey across most articles, not just interactives.

This change removes the rule setting a background colour on the body. It instead sets a background on the article: neutral.100 for most articles, whilst keeping 'transparent' for interactives only.

* Stop using `articleBackgroundLight` for age warning wrapper

It's different to the article background, and using the light background doesn't work when in dark mode.