Update `BigNumber` stories to Component Story Format v3
This is the default for Storybook 7+, and integrates better with some of its features.

Using `satisfies` gives better type safety for args. Replaced the function-based story with the CSFv3 `render` method. Removed the `splitTheme` decorator, as this is now handled by the "global colour scheme" toolbar item and a Chromatic mode. As there's only one story, renamed the story to have the same name as the component, which allows storybook to hoist the story in the UI.