Update `CalloutBlockComponent` Stories To Component Story Format v3 (#12553)
This is the default for Storybook 7+, and integrates better with some of its features.

Using `satisfies` gives better type safety for args. Removed the `splitTheme` decorator, as this is now handled by the "global colour scheme" toolbar item and a Chromatic mode. Added the `centreColumnDecorator`, so the stories better reflect the layout commonly used in articles. Refactored the "good request", "bad request" and "unique id" features into decorators.