Add Datawrapper embed stories
I’ve added a story with a Datawrapper embed for each role, to have an
easy way to look at the styling.

I’ve also tweaked the Wrapper used in the InteractiveBlockComponent
stories to match that used in the Figure stories. Since the layouts are
similar, I think this should be fine?