Separate Datawrapper embed CSS from Figure CSS
To make it clearer which CSS is specific to Datawrapper embeds, this
commit splits the Datawrapper-specific CSS out from the general Figure
CSS, and relies on the ordering of the CSS for having the
Datawrapper-specific CSS override the Figure CSS.

This does mean that there’s a risk the Figure CSS will be edited in a
way that breaks the Datawrapper CSS without the editor realising this.
For example, the Datawrapper CSS uses breakpoints that should be the
same as the ones the Figure CSS uses. Hopefully this risk is outweighed
by the improvement to the clarity of the CSS.