Use dedicated Datawrapper graphic for stories
Because we’re currently using Datawrapper’s “update old versions”
setting, any updates to the Datawrapper graphic used in the stories
cause new UI changes that need to be approved, which is annoying.

To mitigate this, I’ve created a new test graphic which can be kept for
just this purpose, and not updated unless we want there to be changes in
the UI tests.