Create a `Lightbox` component (#12700)
Combines the web and apps lightbox setup into a single component. This puts the logic in a single place, hopefully making it easier to find and understand.

For apps, this also means the component is now referenced once at the `ArticlePage` level, rather than separately in each layout file. This reduces repetition and matches the way it's handled for web.