Move validation out of `enhanceArticleType`
This simplifies the function, as it now just transforms from `FEArticleType` to `Article`. It also means that the article and blocks handlers follow a similar pattern: validate, then enhance.

In addition, it means that it can be used on the "fe-articles" fixtures in tests and stories without having to include the validation step. This will make it easier to rely on the "fe-articles" fixtures only, allowing us to remove the "dcr-articles" fixtures.