Update `CalloutEmbedBlockComponent` stories to CSF v3
Component Story Format v3 is the default for Storybook 7+, and integrates better with some of its features.

Using `satisfies` gives better type safety for args. Applied the "mobileMedium" breakpoint for Chromatic snapshots, as this component doesn't use up the space on the default "wide" breakpoint. Added the `centreColumnDecorator`, so the stories better reflect the layout commonly used in articles. Refactored the "good request" feature into a decorator. As there's only one story, renamed the story to have the same name as the component, which allows storybook to hoist the story in the UI. Also updated the `title` for the related form stories, so that they appear together in the UI.