Fix table element dark mode text (#12810)
The table font colour is switching to an (inherited) light grey in dark mode, but the background colours aren't changing so it clashes. This sets a specific font colour on the table, so that it remains a dark grey on the light table background in both light and dark mode.