Delete the "dcr-articles" fixtures
We currently have both these and the "fe-articles" fixtures. The "fe-articles" fixtures represent the data that's coming into the application, whereas the "dcr-articles" fixtures represent an intermediate data model. This intermediate model can be derived from the "fe-articles" fixtures, but the reverse isn't true. In addition, this model is likely to change more often, and every time it does we have to regenerate the fixtures.

As a result, we'd like to delete the "dcr-articles" fixtures and rely on the "fe-articles" fixtures only from now on.