Move types and functions out of `article.ts` (#12861)
To keep this file focused on the `Article` type.

Moved the `TableOfContentsItem` type to live alongside the functions that create it. Deleted the `TableOfContents` type because it's unused.

Created a `pinnedPost.ts` file to manage enhancing pinned posts alongside the other enhancer files. Updated the type of the `block`/`pinnedPost` field; it's not optional given it's the focus of the function, but rather the value might be `undefined`. Also made some small changes to the implementation for clarity.