Group editions crosswords by date
Updated the shape of the editions crosswords data to be a list of crosswords. Added a `groupByDate` function to divide this up into dates. Added a `CrosswordSelect` component to allow selecting a date and a specific crossword on that date. Added stories and tests for this component.

Added a `Crosswords` component that will call the `groupByDate` function, render the `CrosswordSelect` component, and render the selected crossword itself. This for now just renders the first clue of the selected crossword to demonstrate functionality. Added stories and tests for this component.