Move palette declarations to new file
Any time the `palette` function is imported to get a palette colour, there's a possibility that the whole `palette.ts` module will be included in the bundle. However, the majority of the code in this module is unnecessary if only the `palette` function is needed.

This change moves the palette declarations to a separate file, significantly reducing the size of the `palette` module. From now on the palette declarations should not be included in a bundle when just the `palette` function is needed, and will only appear when the separate `paletteDeclarations` function is directly imported.