Added `StackedProgress` election tracker component (#13007)
This component is one of several that will allow rendering election trackers within DCAR. It's a "stacked progress bar", representing progress through an election divided up by each group running. It's generic, so the kinds of groups it can represent varies. For example:

- Candidates in a US presidential election
- Parties in a UK general election
- Party groups in an EU parliamentary election

These examples are demonstrated in the stories file also included in this change.