Fix submeta scaling (#13066)
The apps include a feature for scaling text and other aspects of the UI by changing the root font size of the page. When this feature is used it causes the fonts in the submeta to scale up and down.

However, the dotted lines that appear between the tag links are created using a background image that relies on precise alignment, calculated in pixels. When the fonts scale this alignment is lost, and the lines start to overlap the text in the tag links.

To fix this, this change migrates several pixel values to their `rem` equivalents. This allows the lengths in the background image, and other aspects of the submeta design, to scale proportionately with the change in root font size.