Refactor quiz atom stories
Added the `centreColumnDecorator`, so the stories better reflect the layout commonly used in articles. Removed the `splitTheme` decorator, as this is now handled by the "global colour scheme" toolbar item and Chromatic modes. Added Chromatic modes to these stories.

Added a `WithResults` story for each variant of the quiz atom. It includes a `play` function to work through the quiz, selecting some answers and revealing/submitting them. This story therefore displays the component in a state that includes the results box, and allows it to be snapshotted like this in Chromatic.