Applied Styles To The Football Match List (#13244)
* Applied styles to the football match list

To match the way it appears on frontend. Updated the stories to display more edge cases: narrow dates, uneven scores. Broke the component down into sub-components to which styles are applied. Added palette light and dark colours.

Also made use of `Intl.DateTimeFormat` to handle formatting the dates and times with appropriate locales and timezones for the different editions. Added functions to the `edition` module to get the timezone and locale for a given edition.

* Move the `grid` module into `src`

So it can be used outside storybook.

* Fix name shadowing in `grid` module

`from` was an import and a function argument.

* Fix name shadowing in `grid` module

`span` was the name of the function and one of its parameters.

---------

Co-authored-by: Marjan Kalanaki <marjan.kalanaki@guardian.co.uk>