Move and rename `TagPage` types
The purpose of this change is to refactor and simplify the codebase,
making it clearer which types and files relate to the data that comes
from frontend. A new `frontend` directory has been created to contain
frontend types and their corresponding schemas. Currently these are
distributed across the codebase and it can be difficult to follow their
usage. The first files to be moved into this directory are
`feTagPage.ts`, a new file to hold the `FETagPage` type, and
`feTagPage.json`, the existing tag page JSON schema file. The other
frontend types and schemas will follow in later changes, as moving them
all together would constitute a very large diff.

This change also renames `FETagPageType` to `FETagPage` and
`DCRTagPageType` to `TagPage`. Removal of the `DCR` prefix has occurred
for two reasons. Firstly, the project is now called DCAR. Secondly,
we're in DCAR (previously DCR), so we can assume all types without a
prefix belong to this project. By contrast, any type that's derived from
frontend remains prefixed with `FE` to differentiate it. The `Type`
suffix has been removed because it's not necessary to suffix a type with
the word `Type`, and this convention was applied inconsistently, so
removing it reduces noise in the codebase.

`FrontPagination` has also been renamed to `TagPagePagination`. Tag
pages are paginated but fronts are not, and this was already partially
reflected by the fact this type was defined in `tagPage.ts` rather than
`front.ts`.