Update `CaptionBlockComponent` stories to CSF v3
Component Story Format v3 is the default for Storybook 7+, and
integrates better with some of its features.

Using `satisfies` gives better type safety for args. Removed the
`splitTheme` decorator, as this is now handled by the "global colour
scheme" toolbar item and a Chromatic mode. Added the
`centreColumnDecorator` to replace the `Wrapper` component. Added a new
`Default` story, and derived the other stories from that. Changed the
Chromatic snapshots to use `mobileMedium`, and added an additional mode
to `WithWidthLimited` to capture the width limiting behaviour at a wider
breakpoint.