Refactor "splitTheme" Decorator (#13422)
Splits up the single "splitTheme" decorator function into multiple
components, functions and types, each with documentation, to make it
easier to understand. Also removes the `theme` arg for two reasons:

1. Similar behaviour is now supported by the `colourSchemeBackground`
and `colourSchemeTextColour` parameters.

2. Configuring addons and decorators is normally achieved using
parameters. The `format` arg is an unusual case, as we want to replace
the story-level `format` arg with the `formats` parameter. In general,
however, using an `arg` carries the risk that we will unintentionally
override any arg in a component that has an identically-named prop. An
example of this is the `EmailSignup` component, which has a `theme` prop
that would be overridden by this decorator.

As a result of this removal, updated the `ShareButton` stories to use
the parameters mentioned in 1).