Update `EmailSignup` Stories To Component Story Format v3 (#13390)
Component Story Format v3 is the default for Storybook 7+, and
integrates better with some of its features.

Using `satisfies` gives better type safety for args. Removed the
`splitTheme` decorator, as this is now handled by the "global colour
scheme" toolbar item and a Chromatic mode. Replaced the now-deprecated
Chromatic `viewports` parameter with Chromatic modes, and added an
additional mode to handle the `wide` breakpoint. Used the `formats`
parameter to set the `ArticleFormat`.