Positioned the football page right column ad
The ad must be positioned in the right column whenever that column is
available. This can be done using the `grid` module, but CSS grid only
works when elements are direct children of the grid container, so the
structure of some of the components has been refactored to make this the
case.

In addition, the `grid` module has been extended to include a
`paddedContainer`, which automatically handles the side margins that are
common to many page designs. This makes it easier to apply borders,
which has also been done in this change.

Co-authored-by: Marjan Kalanaki <marjan.kalanaki@guardian.co.uk>
Co-authored-by: Andrew Howe-Ely <114918544+andrewHEguardian@users.noreply.github.com>