Allow different crops in lightbox
The lightbox code has some de-duplicating logic that is meant to
prevent two or more of the same image appearing in the lightbox. This
was designed to handle cases where galleries repeat the main media image
in the body of the article. Currently the assumption is that two images
are the same if they have the same image id, and can therefore be
de-duped.

However, some articles intentionally display the same image multiple
times, but with different crops. As these will all have the same image
id they will be de-duped, and all but the last image will be dropped
from the lightbox.

The solution is to use both the id and the crop of an image to determine
whether it is unique in the article, so that different crops of the same
image will not be considered identical. However, with this change, if a
gallery uses different crops of the same image in the main media and the
body then these will no longer be considered duplicates, and will both
appear in the lightbox.