Move and rename `Front` types
The purpose of this change is to refactor and simplify the codebase,
making it clearer which types and files relate to the data that comes
from frontend. The new `frontend` directory has been used to contain
frontend types and their corresponding schemas. Currently these are
distributed across the codebase and it can be difficult to follow their
usage. The files in this change to be moved into this directory are
`feFront.ts`, a new file to hold the `FEFront` type, and `feFront.json`,
the existing front JSON schema file. The other frontend types and
schemas will follow in later changes, as moving them all together would
constitute a very large diff.

This change also renames `FEFrontType` to `FEFront` and `DCRFrontType`
to `Front`. Removal of the `DCR` prefix has occurred for two reasons.
Firstly, the project is now called DCAR. Secondly, we're in DCAR
(previously DCR), so we can assume all types without a prefix belong to
this project. By contrast, any type that's derived from frontend remains
prefixed with `FE` to differentiate it. The `Type` suffix has been
removed because it's not necessary to suffix a type with the word
`Type`, and this convention was applied inconsistently, so removing it
reduces noise in the codebase.

The removal of the `Type` suffix has also been applied to any types that
now live in the new `feFront.ts` file. However, the `DCR` prefix and
`Type` suffix remain on any types that still live in the `front.ts`
file. These types are used more widely, so a rename would have resulted
in changes to many more files. They will instead be renamed in a
subsequent change, to avoid making this change too large.