Parse football matches from frontend data (#13463)
Some of the football data pages (live, fixtures, results) render a list
of matches. The model we use to represent these matches in DCAR is
different to that sent as JSON from frontend. For example, it includes
types like `Date` that do not exist as JSON primitives, and it
structures the data in a way that primarily makes sense for rendering in
DCAR but not necessarily for the agents in frontend.

This change adds a parser to transform from the frontend data to DCAR's
internal model. It also generates errors when the data received is not
structured as expected.

In addition, there's a small change to the `Competition` type, renaming
`competitionId` to `id`. It's already a field of the `Competition` type
and therefore does not need the `competition` prefix.

Note that while this parsing approach works, we may want to refine it in
the future by abstracting some of the generic parsing code into another
module or using a library.