Remove lightbox deduplication logic
The lightbox code has some de-duplicating logic that is meant to
prevent two or more of the same image appearing in the lightbox. This
was designed to handle cases where galleries repeat the main media image
in the body of the article. Currently the assumption is that two images
are the same if they have the same image id, and can therefore be
de-duped.

However, some articles intentionally display the same image multiple
times, but with different crops. As these will all have the same image
id they will be de-duped, and all but the last image will be dropped
from the lightbox.

One solution is to use both the id and the crop of an image to determine
whether it is unique in the article, so that different crops of the same
image will not be considered identical. However, galleries often use
different crops of the same image in the main media and the body. As a
result these will no longer be considered duplicates, and will both
appear in the lightbox anyway, defeating the purpose of the
deduplication logic.

The choice, therefore, is between:

1. Two similar images appearing in the lightbox in some galleries.
2. Articles that rely on different crops of the same image having
multiple images missing from the lightbox.

2) is considered more of a problem than 1), so this change removes the
lightbox deduplication logic.