Update `CardCommentCount` stories to CSF v3 (#13621)
Component Story Format v3 is the default for Storybook 7+, and
integrates better with some of its features.

Using `satisfies` gives better type safety for args. Replaced the
`Wrapper` component with a decorator. Since both stories were identical,
deleted the second one. As there was then only one story, renamed the
story to have the same name as the component, which allows storybook to
hoist the story in the UI.