Update Emotion libraries to latest
There have been some changes to Emotion's types which mean that the
various Emotion libraries should be updated together to make sure they
remain in sync.

In addition, these changes have altered the way `EmotionJSX.Element` is
exported. There are two places where this type is used in the codebase,
but elsewhere we generally use `ReactNode` for the same purpose.
Therefore the solution taken here is to switch to that type instead.