Move and rename `Article` types (#13703)
The purpose of this change is to refactor and simplify the codebase,
making it clearer which types and files relate to the data that comes
from frontend. The new `frontend` directory has been used to contain
frontend types and their corresponding schemas. Currently these are
distributed across the codebase and it can be difficult to follow their
usage. The files in this change to be moved into this directory are
`feArticle.ts`, a new file to hold the `FEArticle` type, and
`feArticle.json`, the existing article JSON schema file. The other
frontend types and schemas will follow in later changes, as moving them
all together would constitute a very large diff.

This change also renames `FEArticleType` to `FEArticle`. The `Type`
suffix has been removed because it's not necessary to suffix a type with
the word `Type`, and this convention was applied inconsistently, so
removing it reduces noise in the codebase.