Narrow `dotcom-platform` in CODEOWNERS
When a person or team is marked as a code owner for an area of the
codebase, they are automatically requested for review whenever a PR
modifies that area. At the moment, dotcom-platform is marked as a code
owner for almost the entire codebase. This means that they receive a
review request for essentially every PR that's opened.

This change alters the dotcom-platform team's entry in the CODEOWNERS
file to target specific areas of the codebase. Note that this does not
prevent PR authors from requesting reviews from dotcom-platform, or
anyone else, as needed. It just narrows the list of PRs for which the
team is asked for review automatically.