Use express router in dev server
DCAR uses webpack's built-in dev server for development; when running
`make dev` this is the server that's started up to listen on port 3030.
Its main purpose is to handle compilation and reloading for client-side
assets. However, DCAR is primarily a server-side application, and so
similar functionality for server-side code is useful.

To achieve this, DCAR extends the webpack dev server. It's an instance
of an express app, which can be customised with `setupMiddlewares`, so
we use this to add extra routes and middlewares. To achieve automatic
re-compilation and live reloading of the browser, we rely on the
`webpack-hot-server-middleware` library. Any code for which we require
these features must be built via this library, and must be implemented
as express middleware.

Due to these constraints we don't have access to the original express
app object and its associated routing functionality, and the dev server
is using a custom routing implementation based on a `switch` statement.
However, express has a concept of a `router`, which both supports much
of the functionality of the core app object, such as routing, and also
behaves like middleware. Therefore we can re-write the dev server as a
router, allowing us to use express's built-in routing directly, and
mount further middleware.

This means we no longer have to implement custom routing, and it should
make it easier to extend the dev server with more features in future.
Furthermore, it allows us to mount `getContentFromURLMiddleware` only on
the routes for which it's needed, and get both re-compilation and live
reloading whenever we make changes to it. It will also allow us to
re-write it in TypeScript in the future.